fix(frontend): flatten chapter nav entries
