+++ frontend/tests/components/AlbumNav.test.ts
@@ -41,15 +41,13 @@ const NavChapterGroupStub = defineComponent({
         :key="item.key"
         :data-nav-section="item.key"
       />
-      <template v-for="country in group.countries" :key="country.key">
-        <div
-          v-for="entry in country.entries"
-          :key="entry.type === 'step' ? entry.item.id : entry.key"
-          :data-nav-step="entry.type === 'step' ? entry.item.id : undefined"
-          :data-nav-section="entry.type === 'map' ? entry.key : undefined"
-          :data-range="entry.type === 'map' ? entry.dateRange.join('|') : undefined"
-        />
-      </template>
+      <div
+        v-for="entry in group.entries"
+        :key="entry.type === 'step' ? entry.item.id : entry.key"
+        :data-nav-step="entry.type === 'step' ? entry.item.id : undefined"
+        :data-nav-section="entry.type === 'map' ? entry.key : undefined"
+        :data-range="entry.type === 'map' ? entry.dateRange.join('|') : undefined"
+      />
     </div>`,
 });
